extract updateOrigin
